# release-24.1: sql: avoid unnecessary singleflight requests for role membership cache