--- a/pkg/sql/logictest/testdata/logic_test/crdb_internal_default_privileges
+++ b/pkg/sql/logictest/testdata/logic_test/crdb_internal_default_privileges
@@ -1131,6 +1131,10 @@ ALTER DEFAULT PRIVILEGES IN SCHEMA public GRANT INSERT ON TABLES TO roach_a;
 statement ok
 DELETE FROM system.users WHERE username = 'roach_a';
 
+# Force the role membership cache to be updated.
+statement ok
+CREATE ROLE tmp;
+
 query TTTBTTTB colnames
 SELECT * FROM crdb_internal.default_privileges WHERE grantee = 'roach_a'
 ORDER BY role;