… docs/generated/settings/settings-for-tenants.txt
… docs/generated/settings/settings.html
… pkg/bench/rttanalysis/testdata/benchmark_expectations
… pkg/sql/information_schema.go
… pkg/sql/pg_catalog.go
… pkg/sql/user_test.go